requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cycler==0.11.0
dnspython==2.3.0
Flask==2.2.5
fonttools==4.38.0
gunicorn==22.0.0
idna==3.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.5.3
numpy==1.21.6
packaging==24.0
Pillow==9.5.0
pymongo==4.7.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
requests==2.31.0
six==1.16.0
typing-extensions==4.7.1
urllib3==1.25.11
Werkzeug==2.2.3
zipp==3.15.0
